refactor(ui): modularize agent handling and controls
