+++ tests/test_learning_rate_control.js
@@ -2,10 +2,10 @@ import fs from 'fs';
 import { JSDOM } from 'jsdom';
 
 export async function run(assert) {
-  const js = fs.readFileSync('src/ui/index.js', 'utf8');
+  const js = fs.readFileSync('src/ui/bindControls.js', 'utf8');
   assert.ok(js.includes("const learningRateSlider = document.getElementById('learning-rate-slider');"));
   assert.ok(js.includes("learningRateSlider.addEventListener('input'"));
-  assert.ok(js.includes('agent.learningRate'));
+  assert.ok(js.includes('currentAgent.learningRate'));
   assert.ok(js.includes('learningRateValue.textContent = val.toFixed(2);'));
 
   const dom = new JSDOM(`<input id="learning-rate-slider"><span id="learning-rate-value"></span>`);
